backend-gpu/requirements-worker.txt
onnxruntime-gpu==1.16.3
numpy==1.24.3
redis==5.0.1
Pillow==10.1.0
